% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_schema.R
\name{check_schema}
\alias{check_schema}
\title{Validate a data.frame against a schema}
\usage{
check_schema(data, schema)
}
\arguments{
\item{data}{A data.frame to check}

\item{schema}{A Schema object created with `schema()`}
}
\value{
invisible if validation passes, otherwise stops with error
}
\description{
Validate a data.frame against a schema
}
\examples{

my_schema <- schema(
  mpg ~ is.numeric
)

check_schema(mtcars, my_schema)

}
