% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_incrementing}
\alias{is_incrementing}
\title{Check if the vector is sorted numerically or alphanumerically}
\usage{
is_incrementing(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if the vector is sorted
}
\description{
`NA`s are not ignored and any vector with `NA`s will fail unless the whole vector is `NA`.
}
\examples{
is_incrementing(1:5) # TRUE
is_incrementing(letters[1:5]) # TRUE
is_incrementing(c(4, 3, 0)) # FALSE
}
