% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_non_null}
\alias{is_non_null}
\title{Check if all values are not NA}
\usage{
is_non_null(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if the vector has no NA values
}
\description{
Check if all values are not NA
}
\examples{
is_non_null(1:5) # TRUE
is_non_null(c(1, NA, 3)) # FALSE
}
