% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_positive_integer}
\alias{is_positive_integer}
\title{Check if a vector is a positive integer}
\usage{
is_positive_integer(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if all elements are positive integers (NA ignored)
}
\description{
Check if a vector is a positive integer
}
