% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_text}
\alias{is_text}
\title{Check if a vector is text-based (character or factor)}
\usage{
is_text(x)
}
\arguments{
\item{x}{vector}
}
\value{
TRUE if vector is either character or factor
}
\description{
Check if a vector is text-based (character or factor)
}
