% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_check_errors.R
\name{last_check_errors}
\alias{last_check_errors}
\title{Retrieve latest schematic run time errors}
\usage{
last_check_errors()
}
\value{
error messages
}
\description{
Predicates that error will store the error messages internally and these
can be accessed here.
}
\examples{

last_check_errors()
}
