% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{nullable}
\alias{nullable}
\title{Allow NA in a predicate}
\usage{
nullable(pred)
}
\arguments{
\item{pred}{A predicate function}
}
\value{
A new predicate that allows NAs
}
\description{
Allow NA in a predicate
}
