% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{print.Schema}
\alias{print.Schema}
\title{Print method for Schema}
\usage{
\method{print}{Schema}(x, ...)
}
\arguments{
\item{x}{Object of class Schema}

\item{...}{Other arguments passed to `print()`}
}
\value{
invisible
}
\description{
Print method for Schema
}
