% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\title{Create a schema object}
\usage{
schema(...)
}
\arguments{
\item{...}{Formulae of the form tidyselect_expr ~ predicate}
}
\value{
A Schema object
}
\description{
Create a schema object
}
