% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\title{Create a schema object}
\usage{
schema(..., .names = NULL)
}
\arguments{
\item{...}{Formulae of the form tidyselect_expr ~ predicate}

\item{.names}{Optional vector of names for each of the predicates to be used in the error message in place of the predicate function}
}
\value{
A Schema object
}
\description{
Create a schema object
}
